refactor: decompose lists.py and albums.py into subpackages, fix PDS URL healing
Split two monolithic API files into subpackages following the existing
api/tracks/ pattern:

- lists.py (1149 lines) → lists/{router,schemas,reorder,resolver,playlists}.py
- albums.py (995 lines) → albums/{router,schemas,cache,listing,mutations}.py

Also moves PDS URL healing from lazy per-request side effects (copy-pasted
in 5 API endpoints) to the jetstream identity event handler, where it
belongs. Identity events fire on both handle changes and PDS migrations,
so resolving the DID there keeps the cached pds_url warm proactively
instead of discovering staleness at request time.
