+++ backend/src/backend/api/albums/router.py
@@ -1,0 +1,7 @@
+"""shared FastAPI router for album endpoints."""
+
+from fastapi import APIRouter
+
+router = APIRouter(prefix="/albums", tags=["albums"])
+
+__all__ = ["router"]
