+++ backend/src/backend/api/lists/router.py
@@ -1,0 +1,7 @@
+"""shared FastAPI router for list endpoints."""
+
+from fastapi import APIRouter
+
+router = APIRouter(prefix="/lists", tags=["lists"])
+
+__all__ = ["router"]
