+++ backend/tests/api/test_playlist_by_uri.py
@@ -87,7 +87,7 @@ async def test_get_playlist_by_uri_success(
     """lookup existing playlist by AT-URI returns 200."""
     mock_record = {"value": {"items": []}}
     with patch(
-        "backend.api.lists.get_record_public_resilient",
+        "backend.api.lists.playlists.get_record_public_resilient",
         new_callable=AsyncMock,
         return_value=(mock_record, None),
     ):
@@ -128,7 +128,7 @@ async def test_get_playlist_by_uri_pds_record_not_found(
 ) -> None:
     """playlist exists in DB but PDS record is gone — returns 404 not 500."""
     with patch(
-        "backend.api.lists.get_record_public_resilient",
+        "backend.api.lists.playlists.get_record_public_resilient",
         new_callable=AsyncMock,
         side_effect=RecordNotFound("record not found"),
     ):
@@ -149,7 +149,7 @@ async def test_get_playlist_by_id_pds_record_not_found(
 ) -> None:
     """playlist exists in DB but PDS record is gone — returns 404 not 500."""
     with patch(
-        "backend.api.lists.get_record_public_resilient",
+        "backend.api.lists.playlists.get_record_public_resilient",
         new_callable=AsyncMock,
         side_effect=RecordNotFound("record not found"),
     ):
